Worked on the color scheme
Decided that the colors for the main divs are good. I like that chinese red and
black. still working on text colors. Decided that maybe three different colors
wouldn't be too too overkill. It looks better at any rate. need to work on the
formatting and layout some more. gonna close this sandbox branch out and start
a new one for that.